adjusted tests for backpack items
