+++ Src/DeUrgenta.sln
@@ -28,7 +28,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DeUrgenta.User.Api", "DeUrg
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{34956E68-3F3F-4BFA-935C-9CBE9EFAA352}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeUrgenta.Backpack.Api.Tests", "DeUrgenta.Backpack.Api.Tests\DeUrgenta.Backpack.Api.Tests.csproj", "{88E461A1-6D9D-4980-B6AE-447288F27BA7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeUrgenta.Tests.Helpers", "Tests\DeUrgenta.Tests.Helpers\DeUrgenta.Tests.Helpers.csproj", "{C1B01431-6C78-41A5-B518-658045C77AF0}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DeUrgenta.Backpack.Api.Tests", "Tests\DeUrgenta.Backpack.Api.Tests\DeUrgenta.Backpack.Api.Tests.csproj", "{B133C2DF-EE20-4A85-BEFC-9DA34D519605}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -148,24 +150,37 @@ Global
 		{5A4BBC34-65D6-48E0-BCAD-266FE5E2F376}.Release|x64.Build.0 = Release|Any CPU
 		{5A4BBC34-65D6-48E0-BCAD-266FE5E2F376}.Release|x86.ActiveCfg = Release|Any CPU
 		{5A4BBC34-65D6-48E0-BCAD-266FE5E2F376}.Release|x86.Build.0 = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|x64.Build.0 = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Debug|x86.Build.0 = Debug|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|x64.ActiveCfg = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|x64.Build.0 = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|x86.ActiveCfg = Release|Any CPU
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7}.Release|x86.Build.0 = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|x64.Build.0 = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Debug|x86.Build.0 = Debug|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|x64.ActiveCfg = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|x64.Build.0 = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|x86.ActiveCfg = Release|Any CPU
+		{C1B01431-6C78-41A5-B518-658045C77AF0}.Release|x86.Build.0 = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|x64.Build.0 = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Debug|x86.Build.0 = Debug|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|x64.ActiveCfg = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|x64.Build.0 = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|x86.ActiveCfg = Release|Any CPU
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{88E461A1-6D9D-4980-B6AE-447288F27BA7} = {34956E68-3F3F-4BFA-935C-9CBE9EFAA352}
+		{C1B01431-6C78-41A5-B518-658045C77AF0} = {34956E68-3F3F-4BFA-935C-9CBE9EFAA352}
+		{B133C2DF-EE20-4A85-BEFC-9DA34D519605} = {34956E68-3F3F-4BFA-935C-9CBE9EFAA352}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CB88C361-BFCA-43C0-A7AC-5767ADE33AFC}
