+++ Src/Tests/DeUrgenta.Backpack.Api.Tests/Validators/GetBackpackCategoryItemsValidatorShould.cs
@@ -1,30 +1,36 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 using DeUrgenta.Backpack.Api.Queries;
 using DeUrgenta.Backpack.Api.Validators;
 using DeUrgenta.Domain;
 using DeUrgenta.Domain.Entities;
+using DeUrgenta.Tests.Helpers;
 using Shouldly;
 using Xunit;
 
 namespace DeUrgenta.Backpack.Api.Tests.Validators
 {
-    [Collection("Database collection")]
-    public class GetBackpackItemsValidatorShould
+    [Collection(TestsConstants.DbCollectionName)]
+    public class GetBackpackCategoryItemsValidatorShould
     {
         private readonly DeUrgentaContext _dbContext;
-        public GetBackpackItemsValidatorShould(DatabaseFixture fixture)
+
+        public GetBackpackCategoryItemsValidatorShould(DatabaseFixture fixture)
         {
             _dbContext = fixture.Context;
         }
+
         [Fact]
         public async Task Invalidate_request_when_backpack_does_not_exist()
         {
             // Arrange
-            var sut = new GetBackpackItemsValidator(_dbContext);
+            var sut = new GetBackpackCategoryItemsValidator(_dbContext);
 
             // Act
-            bool isValid = await sut.IsValidAsync(new GetBackpackItems(Guid.NewGuid()));
+            bool isValid = await sut.IsValidAsync(new GetBackpackCategoryItems(Guid.NewGuid(), BackpackCategoryType.WaterAndFood));
 
             // Assert
             isValid.ShouldBeFalse();
@@ -50,10 +56,10 @@ namespace DeUrgenta.Backpack.Api.Tests.Validators
             });
             await _dbContext.SaveChangesAsync();
 
-            var sut = new GetBackpackItemsValidator(_dbContext);
+            var sut = new GetBackpackCategoryItemsValidator(_dbContext);
 
             // Act
-            bool isValid = await sut.IsValidAsync(new GetBackpackItems(backpackId));
+            bool isValid = await sut.IsValidAsync(new GetBackpackCategoryItems(backpackId, BackpackCategoryType.WaterAndFood));
 
             // Assert
             isValid.ShouldBeTrue();
