+++ Src/Tests/DeUrgenta.Backpack.Api.Tests/Validators/GetBackpackItemsValidatorShould.cs
@@ -1,35 +1,31 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using DeUrgenta.Backpack.Api.Queries;
 using DeUrgenta.Backpack.Api.Validators;
 using DeUrgenta.Domain;
 using DeUrgenta.Domain.Entities;
+using DeUrgenta.Tests.Helpers;
 using Shouldly;
 using Xunit;
 
 namespace DeUrgenta.Backpack.Api.Tests.Validators
 {
-    [Collection("Database collection")]
-    public class GetBackpackCategoryItemsValidatorShould
+    [Collection(TestsConstants.DbCollectionName)]
+    public class GetBackpackItemsValidatorShould
     {
         private readonly DeUrgentaContext _dbContext;
-
-        public GetBackpackCategoryItemsValidatorShould(DatabaseFixture fixture)
+        public GetBackpackItemsValidatorShould(DatabaseFixture fixture)
         {
             _dbContext = fixture.Context;
         }
-
         [Fact]
         public async Task Invalidate_request_when_backpack_does_not_exist()
         {
             // Arrange
-            var sut = new GetBackpackCategoryItemsValidator(_dbContext);
+            var sut = new GetBackpackItemsValidator(_dbContext);
 
             // Act
-            bool isValid = await sut.IsValidAsync(new GetBackpackCategoryItems(Guid.NewGuid(), BackpackCategoryType.WaterAndFood));
+            bool isValid = await sut.IsValidAsync(new GetBackpackItems(Guid.NewGuid()));
 
             // Assert
             isValid.ShouldBeFalse();
@@ -55,10 +51,10 @@ namespace DeUrgenta.Backpack.Api.Tests.Validators
             });
             await _dbContext.SaveChangesAsync();
 
-            var sut = new GetBackpackCategoryItemsValidator(_dbContext);
+            var sut = new GetBackpackItemsValidator(_dbContext);
 
             // Act
-            bool isValid = await sut.IsValidAsync(new GetBackpackCategoryItems(backpackId, BackpackCategoryType.WaterAndFood));
+            bool isValid = await sut.IsValidAsync(new GetBackpackItems(backpackId));
 
             // Assert
             isValid.ShouldBeTrue();
