+++ Src/Tests/DeUrgenta.Tests.Helpers/TestsConstants.cs
@@ -1,0 +1,7 @@
+﻿namespace DeUrgenta.Tests.Helpers
+{
+    public class TestsConstants
+    {
+        public const string DbCollectionName = "Database collection";
+    }
+}
